+++ index.html
@@ -231,7 +231,7 @@
     </div>
   </template>
 
-  <script src="js/main.js"></script>
-  
+  <script type="module" src="js/main.js"></script>
+
 </body>
 </html>
